Few minor fixes: fflush for writing to a file.

+++ include/common.h
@@ -138,7 +138,7 @@ int send_echo_request(const int sockfd, struct sockaddr *addr, char *grp_name);
 void start_oneshot_timer(timer_t *t, uint8_t interval, uint32_t sigval);
 void start_recurring_timer(timer_t *t, uint8_t interval, uint32_t sigval);
 unsigned int generate_random_capability(void);
-unsigned int get_task_count(const char*, unsigned int** );
+unsigned int get_task_count(const char*, unsigned long** );
 char * fetch_file(char *, char *);
 void inline create_folder(char * path);
 void inline check_and_create_folder(char * path);
